Move all SDFG debug pass in own file
